Initial support for parsing calendar_dates.txt
I was trying to update the GTFS file to the latest, but it was choking
on special Giants games trains.  So add enough support to add these
special service exceptions to get it to not error out. Ideally this
would also parse the service removal dates for the other schedules, but
that's punted for now.