Key stops by station name instead of entire tuple
stops needs to be keyed by the logical stop. This used to be the (name,
zone) to cross physical stops, but was broken by adding the lat/long
which is different per physical stop. So this reverts back to the
previous behavior to try to match the logical stop by name. This is a
bit different than before, but hopefully station names are unique enough
to not need the zone too.